refactor(tui::ui::components::popups): move Podcast related popups to own file
